requirements.txt
# KeralaSeva AI – Python dependencies
# Install: pip install -r requirements.txt

flask==3.0.3
flask-cors==4.0.1
supabase==2.5.0
python-dotenv==1.0.1
APScheduler==3.10.4
gunicorn==22.0.0
openai